move splash to a more flat file structure, dynamically load reducers
