+++ src/views/splash/activity.js
@@ -1,11 +1,11 @@
 var keyMirror = require('keymirror');
-var api = require('../mixins/api.jsx').api;
+var api = require('../../mixins/api.jsx').api;
 var defaultsDeep = require('lodash.defaultsdeep');
 
 var Types = keyMirror({
-    SET_NEWS: null,
-    SET_NEWS_STATUS: null,
-    SET_NEWS_ERROR: null
+    SET_ACTIVITY: null,
+    SET_ACTIVITY_STATUS: null,
+    SET_ACTIVITY_ERROR: null
 });
 
 module.exports.Status = keyMirror({
@@ -15,20 +15,20 @@ module.exports.Status = keyMirror({
 });
 
 module.exports.getInitialState = function () {
-    return {status: module.exports.Status.FETCHING, news: []};
+    return {status: module.exports.Status.NOT_FETCHED, activity: []};
 };
 
-module.exports.newsReducer = function (state, action) {
+module.exports.activityReducer = function (state, action) {
     // Reducer for handling changes to session state
     if (typeof state === 'undefined') {
         state = module.exports.getInitialState();
     }
     switch (action.type) {
-    case Types.SET_NEWS:
-        return defaultsDeep({news: action.news}, state);
-    case Types.SET_NEWS_STATUS:
+    case Types.SET_ACTIVITY:
+        return defaultsDeep({activity: action.activity}, state);
+    case Types.SET_ACTIVITY_STATUS:
         return defaultsDeep({status: action.status}, state);
-    case Types.SET_NEWS_ERROR:
+    case Types.SET_ACTIVITY_ERROR:
         // TODO: do something with action.error
         return state;
     default:
@@ -36,37 +36,37 @@ module.exports.newsReducer = function (state, action) {
     }
 };
 
-module.exports.setNewsError = function (error) {
+module.exports.setActivityError = function (error) {
     return {
-        type: Types.SET_NEWS_ERROR,
+        type: Types.SET_ACTIVITY_ERROR,
         error: error
     };
 };
 
-module.exports.setNews = function (news) {
+module.exports.setActivity = function (activity) {
     return {
-        type: Types.SET_NEWS,
-        news: news
+        type: Types.SET_ACTIVITY,
+        activity: activity
     };
 };
 
 module.exports.setStatus = function (status){
     return {
-        type: Types.SET_NEWS_STATUS,
+        type: Types.SET_ACTIVITY_STATUS,
         status: status
     };
 };
 
-module.exports.getNews = function () {
+module.exports.getActivity = function (username) {
     return function (dispatch) {
-        dispatch(module.exports.setStatus(module.exports.Status.NOT_FETCHED));
+        dispatch(module.exports.setStatus(module.exports.Status.FETCHING));
         api({
-            uri: '/news?limit=3'
+            uri: '/proxy/users/' + username + '/activity?limit=5'
         }, function (err, body) {
-            if (err) return dispatch(module.exports.setNewsError(err));
+            if (err) return dispatch(module.exports.setActivityError(err));
 
             if (typeof body !== 'undefined') {
-                dispatch(module.exports.setNews(body));
+                dispatch(module.exports.setActivity(body));
                 dispatch(module.exports.setStatus(module.exports.Status.FETCHED));
                 return;
             }
