--- src/views/splash/components/splash.scss
@@ -1,37 +1,0 @@
-.splash {
-    .splash-header {
-        display: flex;
-        flex-direction: row;
-        flex-wrap: nowrap;
-        justify-content: space-between;
-
-        .ie9 & {
-            display: table;
-            margin: 0 -20px 20px -20px;
-            min-width: 100%;
-            border-spacing: 20px 0;
-        }
-
-        .box {
-            display: inline-block;
-            width: calc(60% - 20px);
-
-            .ie9 & {
-                display: table-cell;
-                vertical-align: top;
-            }
-        }
-
-        .news {
-            width: 40%;
-        }
-    }
-
-    .teacher-banner {
-        margin-top: -20px;
-    }
-
-    .box {
-        margin-bottom: 20px;
-    }
-}
