+++ src/views/splash/news.js
@@ -1,11 +1,11 @@
 var keyMirror = require('keymirror');
-var api = require('../mixins/api.jsx').api;
+var api = require('../../mixins/api.jsx').api;
 var defaultsDeep = require('lodash.defaultsdeep');
 
 var Types = keyMirror({
-    SET_ACTIVITY: null,
-    SET_ACTIVITY_STATUS: null,
-    SET_ACTIVITY_ERROR: null
+    SET_NEWS: null,
+    SET_NEWS_STATUS: null,
+    SET_NEWS_ERROR: null
 });
 
 module.exports.Status = keyMirror({
@@ -15,20 +15,20 @@ module.exports.Status = keyMirror({
 });
 
 module.exports.getInitialState = function () {
-    return {status: module.exports.Status.NOT_FETCHED, activity: []};
+    return {status: module.exports.Status.FETCHING, news: []};
 };
 
-module.exports.activityReducer = function (state, action) {
+module.exports.newsReducer = function (state, action) {
     // Reducer for handling changes to session state
     if (typeof state === 'undefined') {
         state = module.exports.getInitialState();
     }
     switch (action.type) {
-    case Types.SET_ACTIVITY:
-        return defaultsDeep({activity: action.activity}, state);
-    case Types.SET_ACTIVITY_STATUS:
+    case Types.SET_NEWS:
+        return defaultsDeep({news: action.news}, state);
+    case Types.SET_NEWS_STATUS:
         return defaultsDeep({status: action.status}, state);
-    case Types.SET_ACTIVITY_ERROR:
+    case Types.SET_NEWS_ERROR:
         // TODO: do something with action.error
         return state;
     default:
@@ -36,37 +36,37 @@ module.exports.activityReducer = function (state, action) {
     }
 };
 
-module.exports.setActivityError = function (error) {
+module.exports.setNewsError = function (error) {
     return {
-        type: Types.SET_ACTIVITY_ERROR,
+        type: Types.SET_NEWS_ERROR,
         error: error
     };
 };
 
-module.exports.setActivity = function (activity) {
+module.exports.setNews = function (news) {
     return {
-        type: Types.SET_ACTIVITY,
-        activity: activity
+        type: Types.SET_NEWS,
+        news: news
     };
 };
 
 module.exports.setStatus = function (status){
     return {
-        type: Types.SET_ACTIVITY_STATUS,
+        type: Types.SET_NEWS_STATUS,
         status: status
     };
 };
 
-module.exports.getActivity = function (username) {
+module.exports.getNews = function () {
     return function (dispatch) {
-        dispatch(module.exports.setStatus(module.exports.Status.FETCHING));
+        dispatch(module.exports.setStatus(module.exports.Status.NOT_FETCHED));
         api({
-            uri: '/proxy/users/' + username + '/activity?limit=5'
+            uri: '/news?limit=3'
         }, function (err, body) {
-            if (err) return dispatch(module.exports.setActivityError(err));
+            if (err) return dispatch(module.exports.setNewsError(err));
 
             if (typeof body !== 'undefined') {
-                dispatch(module.exports.setActivity(body));
+                dispatch(module.exports.setNews(body));
                 dispatch(module.exports.setStatus(module.exports.Status.FETCHED));
                 return;
             }
