+++ src/views/splash/splash.scss
@@ -1,3 +1,5 @@
+@import "../../colors";
+
 .splash {
     .splash-header {
         display: flex;
@@ -29,9 +31,43 @@
 
     .teacher-banner {
         margin-top: -20px;
+        background-color: $ui-purple;
+        min-height: 65px;
+
+        &.title-banner {
+            transition: none;
+            margin-bottom: 20px;
+            text-align: left;
+
+            h3,
+            p {
+                margin: 0;
+                padding: 0;
+                width: 100%;
+                text-align: left;
+            }
+        }
+
+        h3 {
+            color: $ui-white;
+        }
+
+        .flex-row {
+            &.inner {
+                justify-content: space-between;
+            }
+        }
+
+        .button {
+            margin-left: 10px;
+            background-color: $ui-white;
+            padding: 13px 20px;
+            color: $ui-blue;
+        }
     }
+
 
     .box {
         margin-bottom: 20px;
     }
-}    
+}
